Web-project: edit models & add all DAO implementations
* Rewrite hashCode() and equals(Object obj) fom each model;
* Add implementations for all DAO interfaces;

+++ Web-project/dataaccess/src/main/java/com/github/pixelase/dataaccess/dao/BrigadeDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.github.pixelase.dataaccess.model.Brigade;
 
-public interface BrigadeDao extends PagingAndSortingRepository<Brigade, Long> {
+public interface BrigadeDao extends PagingAndSortingRepository<Brigade, Integer> {
 
 }
