+++ Web-project/dataaccess/src/main/java/com/github/pixelase/dataaccess/dao/EmployeeDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.github.pixelase.dataaccess.model.Employee;
 
-public interface EmployeeDao extends PagingAndSortingRepository<Employee, Long> {
+public interface EmployeeDao extends PagingAndSortingRepository<Employee, Integer> {
 
 }
