+++ Web-project/dataaccess/src/main/java/com/github/pixelase/dataaccess/dao/WorkScopeDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.github.pixelase.dataaccess.model.WorkScope;
 
-public interface WorkScopeDao extends PagingAndSortingRepository<WorkScope, Long> {
+public interface WorkScopeDao extends PagingAndSortingRepository<WorkScope, Integer> {
 
 }
