+++ Web-project/dataaccess/src/main/java/com/github/pixelase/dataaccess/dao/WorkTypeDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.github.pixelase.dataaccess.model.WorkType;
 
-public interface WorkTypeDao extends PagingAndSortingRepository<WorkType, Long> {
+public interface WorkTypeDao extends PagingAndSortingRepository<WorkType, Integer> {
 
 }
